Homework_8

+++ src/Figures/Rectangle.java
@@ -1,18 +1,18 @@
 package Figures;
 
 public class Rectangle extends Figure {
-    public Rectangle(String figureType, int a, int b) {
+    public Rectangle(String figureType, double a, double b) {
         super(figureType);
         this.a = a;
         this.b = b;
     }
 
-    int a;
-    int b;
+    double a;
+    double b;
 
     @Override
     double calculateArea() {
-        int s = a * b;
+        double s = a * b;
         return s;
     }
 
@@ -22,45 +22,3 @@ public class Rectangle extends Figure {
         return p;
     }
 }
-//public class Rectangle extends Figure {
-//    private static final String NAME = "Rectangle";
-//
-//    private double width;
-//    private double height;
-//
-//    public Rectangle(double width, double height) {
-//        this.width = width;
-//        this.height = height;
-//    }
-//
-//    @Override
-//    public double getArea() {
-//        return width * height;
-//    }
-//    @Override
-//    public double getPerimeter() {
-//        double perimeter = (width + height) * 2;
-//        return perimeter;
-//    }
-//
-//    @Override
-//    public String getName() {
-//        return NAME;
-//    }
-//
-//    public double getWidth() {
-//        return width;
-//    }
-//
-//    public void setWidth(double width) {
-//        this.width = width;
-//    }
-//
-//    public double getHeight() {
-//        return height;
-//    }
-//
-//    public void setHeight(double height) {
-//        this.height = height;
-//    }
-//}
